+++ .bash_history
@@ -1,0 +1,47 @@
+ls
+vim CMakeLists.txt 
+vim Dockerfile 
+mv /usr/local/share/mingw32/ /usr/local/share/CUnit
+clear
+vim CMakeLists.txt 
+ls 
+vim CMakeLists.txt 
+fg
+fg
+make
+vim CMakeLists.txt 
+make
+vim Cm
+vim CMakeLists.txt 
+make
+vim Makefile.old 
+vim CMakeLists.txt 
+make
+vim Makefile.old 
+make -f Makefile.old 
+make -f Makefile.old run-tests
+exit
+make
+make extra-clean
+ls
+vim CMakeLists.txt 
+make
+clear
+ls
+exit
+ls /usr/local/share/
+make
+vim CMakeLists.txt 
+make
+vim Makefile.old 
+vim CMakeLists.txt 
+make
+make ex
+make extra-clean
+vim CMakeLists.txt 
+make
+vim Makefile.old 
+make -f Makefile.old run-tests
+exit
+make
+exit
