+++ README.md
@@ -1,4 +1,4 @@
-<img src="./resources/Banner.png"></img>
+<img src="./res/Banner.png"></img>
 
 **Pidjeon**, a mash of the words PID (process ID) and injection, is a command  
 line application for Windows that uses various techniques to inject a specified  
@@ -9,18 +9,19 @@ payload into a target process.
 ## Demonstration
 [ embed a demo ]
 
-## Environment (Requires Docker)
+## Load Docker Environment (Requires Docker)
 ```bash
+# using WSL2
 git clone https://github.com/0xvpr/Pidjeon
 cd Pidjeon
-chmod +x ./env/docker-buld.sh ./env/docker-run.sh
-./env/docker-buld.sh && ./env/docker-run.sh
+make docker-container && make-instance
 ```
 
 ## Usage
 ### Primary Features
 - [x] DLL injection
 - [x] Remote thread injection & execution
+- [x] Shellcode injection
 
 ### Switches/Usage
 - SWITCH: -i [ LoadLibraryA | LoadLibraryW | ManualMap ] -\> specify the injection method
@@ -28,6 +29,8 @@ chmod +x ./env/docker-buld.sh ./env/docker-run.sh
 - SWITCH: -o --output \<path/to/output\> -\> specify log file location
 - SWITCH: -r --remote -> specify remote thread routine
 - SWITCH: -v -> specify verbosity level
+- SWITCH: -c -> create a remote thread using second positional argument
+
 
 ### Examples
 #### Powershell
