+++ TODO.md
@@ -1,0 +1,4 @@
+# TODO
+- [ ] Use automake build system
+  - [ ] Use cmake build system inside of pidjeon
+- [ ] IDK
