+++ mingw-toolchain.cmake
@@ -1,5 +1,9 @@
-# Set compilers
-set(CMAKE_C_COMPILER i686-w64-mingw32-gcc-posix)
+## Set compilers
+if     ( "${PROJECT_ARCHITECTURE}" STREQUAL "x86" )
+    set( CMAKE_C_COMPILER "i686-w64-mingw32-gcc-posix" )
+elseif ( "${PROJECT_ARCHITECTURE}" STREQUAL "x64" )
+    set( CMAKE_C_COMPILER "x86_64-w64-mingw32-gcc-posix" )
+endif()
 
-# Set compile flags
-set(CMAKE_C_FLAGS "-std=c17 -O2 -s -Wall -Wextra -Wshadow -Wpedantic -Wconversion -ffast-math -fvisibility=hidden -funroll-loops -fPIE")
+## Set compile flags
+set( CMAKE_C_FLAGS "-std=c99 -O2 -s -Wall -Wextra -Wshadow -Wpedantic -Wconversion -fno-ident -ffast-math -fvisibility=hidden -funroll-loops -fPIE" )
