+++ src/dummy/compile_flags.txt
@@ -1,0 +1,13 @@
+-std=c99
+--target=x86_64-pc-windows-gnu
+-Iinclude
+-isystem/usr/x86_64-w64-mingw32/include
+-isystem/usr/lib/gcc/x86_64-w64-mingw32/9.3-win32/include
+-isystem/usr/lib/gcc/x86_64-w64-mingw32/9.3-win32/include-fixed
+-Wall
+-Wextra
+-Werror
+-Wshadow
+-Wpedantic
+-Wconversion
+-Wno-pointer-to-int-cast
